Speed up assertEqualTextFiles
The issue was that XReadLines returned a long LinkedList on which we then did
random access. These are slow on linked lists.

This commit fixes XReadLines so it returns an ArrayList instead:
this has the same asymptotic performance but will work much better
for callers that then try to do random access.

This commit also fixes assertEqualTextFiles's performance. There was no
reason to slurp in the whole files before doing the comparison: the new
code instead does the comparisons as we read the files. This reduces
memory usage as we no longer hold the whole files in memory, and as a
bonus should perform a little bit faster.

VariantRecalibratorIntegrationTest for example was affected by this
issue: it went from 1m29s normally to 25m50s when using XReadLines
(Travis timings). With this commit, VariantRecalibratorIntegrationTest
now takes 1m4s (on my machine). We can't compare directly because these
are different machines, but it at least strongly suggests that the
performance regression has been fixed.